StepHandler and StepPause internal structures added [#22]
