+++ maiko/src/internal.rs
@@ -1,7 +1,11 @@
 mod actor_handler;
 mod broker;
+mod step_handler;
+mod step_pause;
 mod subscriber;
 
 pub(crate) use actor_handler::ActorHandler;
 pub(crate) use broker::Broker;
+pub(crate) use step_handler::StepHandler;
+pub(crate) use step_pause::StepPause;
 pub(crate) use subscriber::Subscriber;
