+++ maiko/src/internal/step_pause.rs
@@ -1,0 +1,7 @@
+#[derive(Debug, Default, Clone, Copy, PartialEq, Eq)]
+pub(crate) enum StepPause {
+    #[default]
+    None,
+    AwaitEvent,
+    Suppressed,
+}
